+++ packages/tasit-action/src/erc20/Mana.js
@@ -1,7 +1,7 @@
 import Contract from "./Contract";
 import manaTokenABI from "./abi/MANAToken.json";
 
-export default class DecentralandMana extends Contract {
+export default class Mana extends Contract {
   constructor(address, wallet) {
     const abi = manaTokenABI;
     super(address, abi, wallet);
