+++ packages/tasit-action/src/erc20/Mana.test.js
@@ -1,11 +1,11 @@
-import DecentralandMana from "./DecentralandMana";
+import Mana from "./Mana";
 
 import { local as localAddresses } from "../../tasit-contracts/decentraland/addresses";
 const { MANAToken: MANA_ADDRESS } = localAddresses;
 
-describe("TasitAction.DecentralandMana", () => {
+describe("TasitAction.Decentraland.Mana", () => {
   it("should get the MANAToken name", async function() {
-    const mana = new DecentralandMana(MANA_ADDRESS);
+    const mana = new Mana(MANA_ADDRESS);
     const name = await mana.name();
     expect(name).to.equal("Decentraland MANA");
   });
