+++ packages/tasit-action/src/erc721/Land.js
@@ -1,7 +1,7 @@
 import Contract from "./Contract";
 import landRegistryABI from "./abi/LANDRegistry.json";
 
-export default class DecentralandLand extends Contract {
+export default class Land extends Contract {
   constructor(address, wallet) {
     const abi = landRegistryABI;
     super(address, abi, wallet);
