+++ packages/tasit-action/src/marketplace/Decentraland.js
@@ -1,7 +1,7 @@
 import Contract from "./Contract";
 import marketplaceABI from "./abi/Marketplace.json";
 
-export default class DecentralandMarketplace extends Contract {
+export default class Decentraland extends Contract {
   constructor(address, wallet) {
     const abi = marketplaceABI;
     super(address, abi, wallet);
