+++ packages/tasit-action/src/marketplace/Decentraland.test.js
@@ -1,11 +1,11 @@
-import DecentralandMarketplace from "./DecentralandMarketplace";
+import Decentraland from "./Decentraland";
 
 import { local as localAddresses } from "../../tasit-contracts/decentraland/addresses";
 const { Marketplace: MARKETPLACE_ADDRESS } = localAddresses;
 
-describe("TasitAction.DecentralandMarketplace", () => {
+describe("TasitAction.Marketplace.Decentraland", () => {
   it("contract shouldn't be paused", async function() {
-    const marketplace = new DecentralandMarketplace(MARKETPLACE_ADDRESS);
+    const marketplace = new Decentraland(MARKETPLACE_ADDRESS);
     const isPaused = await marketplace.paused();
     expect(isPaused).to.equal(false);
   });
